# docs/swagger.yaml
schemes:
  - http
  - https
swagger: '2.0'
info:
  description: REST API for the LMS platform covering auth, courses, enrollments, assessments, payments, and more.
  title: LMS Backend API
  contact: {}
  version: 1.0.0
host: localhost:8080
basePath: /
paths:
...
  /v1/admin/system/settings:
    get:
      security:
        - BearerAuth: []
      description: Returns the current platform settings
      produces:
        - application/json
      tags:
        - system-config
      summary: Get system settings
      responses:
        '200':
          description: OK
          schema:
            $ref: '#/definitions/system_config.SettingsResponse'
        '401':
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        '403':
          description: Forbidden
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
    patch:
      security:
        - BearerAuth: []
      description: Applies a partial update to platform settings
      consumes:
        - application/json
      produces:
        - application/json
      tags:
        - system-config
      summary: Update system settings
      parameters:
        - description: Settings update data
          name: body
          in: body
          required: true
          schema:
            type: object
      responses:
        '200':
          description: OK
          schema:
            $ref: '#/definitions/system_config.SettingsResponse'
        '400':
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ValidationErrorResponse'
        '401':
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        '403':
          description: Forbidden
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
  /v1/admin/system/settings/history:
    get:
      security:
        - BearerAuth: []
      description: Returns paginated settings change history
      produces:
        - application/json
      tags:
        - system-config
      summary: Get settings history
      parameters:
        - type: integer
          default: 1
          description: Page number
          name: page
          in: query
        - type: integer
          default: 20
          description: Items per page
          name: limit
          in: query
      responses:
        '200':
          description: OK
          schema:
            $ref: '#/definitions/system_config.SettingsHistoryResponse'
        '401':
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        '403':
          description: Forbidden
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
  /v1/admin/system/settings/rollback/{historyId}:
    post:
      security:
        - BearerAuth: []
      description: Restores platform settings to a historical snapshot
      produces:
        - application/json
      tags:
        - system-config
      summary: Rollback settings
      parameters:
        - type: string
          description: History entry ID
          name: historyId
          in: path
          required: true
      responses:
        '200':
          description: OK
          schema:
            $ref: '#/definitions/system_config.RollbackResponse'
        '400':
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ValidationErrorResponse'
        '401':
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        '403':
          description: Forbidden
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        '404':
          description: Not Found
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
...
definitions:
...
  handlers.ErrorResponse:
    type: object
    properties:
      error:
        type: object
        properties:
          code:
            type: string
          message:
            type: string
  handlers.ValidationDetail:
    type: object
    properties:
      field:
        type: string
      message:
        type: string
  handlers.ValidationErrorResponse:
    type: object
    properties:
      error:
        type: object
        properties:
          code:
            type: string
          details:
            type: array
            items:
              $ref: '#/definitions/handlers.ValidationDetail'
          message:
            type: string
...
  pagination.Meta:
    type: object
    properties:
      limit:
        type: integer
      page:
        type: integer
      total:
        type: integer
      total_pages:
        type: integer
...
  system_config.RollbackResponse:
    type: object
    properties:
      restored_from_history_id:
        type: string
      settings:
        $ref: '#/definitions/system_config.SettingsResponse'
  system_config.SettingsHistoryItem:
    type: object
    properties:
      changed_at:
        type: string
      changed_by:
        type: string
      diff:
        type: object
      id:
        type: string
      snapshot:
        type: object
  system_config.SettingsHistoryResponse:
    type: object
    properties:
      data:
        type: array
        items:
          $ref: '#/definitions/system_config.SettingsHistoryItem'
      meta:
        $ref: '#/definitions/pagination.Meta'
  system_config.SettingsResponse:
    type: object
    properties:
      default_timezone:
        type: string
      feature_flags:
        type: object
      maintenance_mode:
        type: boolean
      oauth_providers_enabled:
        type: array
        items:
          type: string
      platform_name:
        type: string
      updated_at:
        type: string
      updated_by:
        type: string
securityDefinitions:
  BearerAuth:
    type: apiKey
    name: Authorization
    in: header